tech-studio-projects/financial-industry/algo-trading-platform/python/core-monolith/trading_engine.py
# src/main/python/yourcompany/engine/trading_engine.py

class TradingEngine:
    def __init__(self, execution_service, risk_manager):
        self.execution_service = execution_service
        self.risk_manager = risk_manager

    def execute_trade(self, symbol, quantity, price, order_type):
        """Executes a trade if risk checks pass."""
        if self.risk_manager.check_risk(symbol, quantity):
            return self.execution_service.place_order(symbol, quantity, price, order_type)
        else:
            return {"error": "Risk criteria not met"}

# Example usage:
# execution_service = ExecutionService()  # Placeholder for execution service
# risk_manager = RiskManager()  # Placeholder for risk manager
# engine = TradingEngine(execution_service, risk_manager)
# engine.execute_trade("AAPL", 100, 150.25, "market")
