tech-studio-projects/financial-industry/algo-trading-platform/python/core-monolith/tests/test_trading_engine.py
# tests/python/yourcompany/strategy/test_trading_strategy.py

import pytest
from unittest.mock import MagicMock
from yourcompany.strategy.trading_strategy import TradingStrategy
from yourcompany.engine.trading_engine import TradingEngine
from yourcompany.execution.order_execution import ExecutionService
from yourcompany.risk.risk_manager import RiskManager

@pytest.fixture
def trading_strategy():
    """Fixture for TradingStrategy instance."""
    execution_service = MagicMock(ExecutionService)
    risk_manager = MagicMock(RiskManager)
    trading_engine = TradingEngine(execution_service, risk_manager)
    return TradingStrategy(trading_engine)

def test_run_strategy_execute_trade(trading_strategy):
    """
    Test that the strategy executes a trade when the conditions are met.
    """
    trading_strategy.trading_engine.execute_trade = MagicMock(return_value={"status": "success", "tradeID": "123456"})
    
    result = trading_strategy.run("AAPL", 100, 150.25, "limit")
    
    # Assert that the trade was executed
    assert result["status"] == "success"
    assert result["tradeID"] == "123456"

def test_run_strategy_does_not_execute_trade(trading_strategy):
    """
    Test that the strategy does not execute a trade when conditions are not met.
    """
    trading_strategy.trading_engine.execute_trade = MagicMock(return_value={"error": "Conditions not met"})
    
    result = trading_strategy.run("AAPL", 100, 140.25, "limit")
    
    # Assert that the trade was not executed due to conditions
    assert result["error"] == "Conditions not met"
